Move the decoder to a separate file
